requirements.txt
streamlit==1.35.0
tensorflow==2.13.0
tensorflow==2.8.0
tensorflow==2.12.0
numpy==1.24.3
Pillow==10.2.0
h5py==3.9.0s